+++ tests/test_transforms/test_decomposition/test_model_based.py
@@ -337,3 +337,38 @@ def test_stride_transform(forward_stride_datasets, decompose_model, context_size
 
     assert not transformed.df.iloc[context_size:10].isna().any().any()
     assert transformed.df.iloc[10:].isna().all().any()
+
+
+@pytest.mark.parametrize("features_to_ignore", (["target"], ["regressor_1"], ["target", "regressor_1"]))
+def test_repetitive_call_before_pipeline_with_ignore(outliers_solid_tsds, features_to_ignore):
+    """Problem arises when decomposition transform is applied the second time to the same `TSDataset`.
+    This test mimics behavior when we do analysis before building pipeline and want to apply results in pipeline
+    transforms without copying the dataset.
+
+    In such cases error `KeyError: '[] not found in axis'` may be raised by pandas.
+    """
+    ts = outliers_solid_tsds
+
+    ts.fit_transform(
+        [
+            ModelDecomposeTransform(
+                model=HoltWintersModel(trend="add", seasonal="add", seasonal_periods=None), residuals=True
+            )
+        ]
+    )
+
+    pipeline = Pipeline(
+        transforms=[
+            ModelDecomposeTransform(
+                model=HoltWintersModel(trend="add", seasonal="add", seasonal_periods=None), residuals=True
+            ),
+            IForestOutlierTransform(
+                in_column="target",
+                features_to_ignore=features_to_ignore,
+            ),
+            TimeSeriesImputerTransform(),
+        ],
+        model=SARIMAXModel(),
+    )
+
+    pipeline.fit(ts)